CiviCRM 3.3.1 core upgrade, with minor mods to jquery plugins

+++ modules/civicrm/CRM/Member/BAO/Membership.php
@@ -2,7 +2,7 @@
 
  /*
  +--------------------------------------------------------------------+
- | CiviCRM version 3.2                                                |
+ | CiviCRM version 3.3                                                |
  +--------------------------------------------------------------------+
  | Copyright CiviCRM LLC (c) 2004-2010                                |
  +--------------------------------------------------------------------+
@@ -268,8 +268,8 @@ class CRM_Member_BAO_Membership extends CRM_Member_DAO_Membership
             $contributionParams = array( );
             $config = CRM_Core_Config::singleton();
             $contributionParams['currency'  ] = $config->defaultCurrency;
-            $contributionParams['receipt_date'] = $params['receipt_date'] ? $params['receipt_date'] : 'null';
-            $contributionParams['source']       = $params['contribution_source'];
+            $contributionParams['receipt_date'] = ( CRM_Utils_Array::value('receipt_date', $params) ) ? $params['receipt_date'] : 'null';
+            $contributionParams['source']       = CRM_Utils_Array::value('contribution_source', $params);
             $contributionParams['non_deductible_amount'] = 'null';
             $recordContribution = array( 'contact_id', 'total_amount', 'receive_date', 'contribution_type_id', 'payment_instrument_id', 'trxn_id', 'invoice_id', 'is_test', 'contribution_status_id', 'check_number' );
             foreach ( $recordContribution as $f ) {
@@ -345,13 +345,26 @@ class CRM_Member_BAO_Membership extends CRM_Member_DAO_Membership
             $membershipTypes = CRM_Member_PseudoConstant::membershipType();
             $title = CRM_Contact_BAO_Contact::displayName( $membership->contact_id ) . ' - ' . ts('Membership Type:') . ' ' . $membershipTypes[$membership->membership_type_id];
             
+            require_once 'CRM/Core/Permission.php';
+            $recentOther = array( );
+            if ( CRM_Core_Permission::checkActionPermission( 'CiviMember', CRM_Core_Action::UPDATE ) ) {
+                $recentOther['editUrl'] = CRM_Utils_System::url( 'civicrm/contact/view/membership', 
+                                                                 "action=update&reset=1&id={$membership->id}&cid={$membership->contact_id}&context=home" );
+            } 
+            if ( CRM_Core_Permission::checkActionPermission( 'CiviMember', CRM_Core_Action::DELETE ) ) {
+                $recentOther['deleteUrl'] = CRM_Utils_System::url( 'civicrm/contact/view/membership', 
+                                                                   "action=delete&reset=1&id={$membership->id}&cid={$membership->contact_id}&context=home" );
+            }
+
             // add the recently created Membership
             CRM_Utils_Recent::add( $title,
                                    $url,
                                    $membership->id,
                                    'Membership',
                                    $membership->contact_id,
-                                   null );
+                                   null,
+                                   $recentOther
+                                   );
         }
         
         return $membership;
@@ -397,12 +410,21 @@ class CRM_Member_BAO_Membership extends CRM_Member_DAO_Membership
                 $relType   = array( 'id' => $values['civicrm_relationship_type_id'] );
                 $relValues = array( );
                 CRM_Contact_BAO_RelationshipType::retrieve( $relType, $relValues);
-                
-                // 1. Check if contact and membership type relationship type are same
-                // 2. Check if relationship direction is same or name_a_b = name_b_a
-                if ( ( $values['civicrm_relationship_type_id'] == $membershipType['relationship_type_id'] )
-                     && ( ( $values['rtype'] == $membershipType['relationship_direction'] ) ||
-                          ( $relValues['name_a_b'] == $relValues['name_b_a'] ) ) ) {
+                // Check if contact's relationship type exists in membership type
+                $relTypeDirs   = array( );
+                $relTypeIds    = explode( CRM_Core_DAO::VALUE_SEPARATOR,$membershipType['relationship_type_id'] );
+                $relDirections = explode( CRM_Core_DAO::VALUE_SEPARATOR,$membershipType['relationship_direction'] );
+                $bidirectional = false;
+                foreach ( $relTypeIds as $key => $value ) {
+                    $relTypeDirs[] = $value."_".$relDirections[$key];
+                    if ( in_array( $value, $relType ) &&
+                         $relValues['name_a_b'] == $relValues['name_b_a'] ) {
+                        $bidirectional = true;
+                        break;
+                    }
+                }
+                $relTypeDir = $values['civicrm_relationship_type_id'].'_'.$values['rtype'];
+                if ( $bidirectional || in_array( $relTypeDir, $relTypeDirs ) ) {
                     // $values['status'] is going to have value for
                     // current or past relationships.
                     $contacts[$values['cid']] = $values['status'];
@@ -431,34 +453,55 @@ class CRM_Member_BAO_Membership extends CRM_Member_DAO_Membership
     static function retrieve( &$params, &$defaults ) 
     {
         $membership = new CRM_Member_DAO_Membership( );
+        
         $membership->copyValues( $params );
-        $idList = array('membership_type' => 'MembershipType',
-                        'status'          => 'MembershipStatus',
-                        );
+        
         if ( $membership->find( true ) ) {
             CRM_Core_DAO::storeValues( $membership, $defaults );
-            foreach ( $idList as $name => $file ) {
-                if ( $defaults[$name .'_id'] ) {
-                    $defaults[$name] = CRM_Core_DAO::getFieldValue( 'CRM_Member_DAO_' . $file, 
-                                                                    $defaults[$name .'_id'] );
-                }
+            
+            //get the membership status and type values.
+            $statusANDType = self::getStatusANDTypeVaues( $membership->id );
+            foreach ( array( 'status', 'membership_type' ) as $fld ){
+                $defaults[$fld] = CRM_Utils_Array::value( $fld, $statusANDType[$membership->id] );
             }
-
-            if ( $membership->status_id ) {
-                $active = CRM_Core_DAO::getFieldValue('CRM_Member_DAO_MembershipStatus',
-                                                      $membership->status_id,
-                                                      'is_current_member');
-                if ( $active ) {
-                    $defaults['active'] = $active;
-                }
+            if ( CRM_Utils_Array::value( 'is_current_member', $statusANDType[$membership->id] ) ) {
+                $defaults['active'] = true;
             }
-
+            
             $membership->free( );
+            
             return $membership;
         }
+        
         return null;
     }
-
+    
+    function getStatusANDTypeVaues( $membershipId ) 
+    {
+        $values = array( );
+        if ( !$membershipId ) return $values;
+        $sql = '
+    SELECT  membership.id as id,
+            status.id as status_id,
+            status.label as status,
+            status.is_current_member as is_current_member,
+            type.id as membership_type_id,
+            type.name as membership_type
+      FROM  civicrm_membership membership
+INNER JOIN  civicrm_membership_status status ON ( status.id = membership.status_id )
+INNER JOIN  civicrm_membership_type type ON ( type.id = membership.membership_type_id )
+     WHERE  membership.id = %1';
+        $dao = CRM_Core_DAO::executeQuery( $sql, array( 1 => array( $membershipId, 'Positive' ) ) );
+        $properties = array( 'status', 'status_id', 'membership_type', 'membership_type_id', 'is_current_member' );
+        while ( $dao->fetch( ) ) {
+            foreach ( $properties as $property ) {
+                $values[$dao->id][$property] = $dao->$property;
+            }
+        }
+        
+        return $values;
+    }
+    
     /** 
      * Function to delete membership.
      * 
@@ -749,9 +792,9 @@ class CRM_Member_BAO_Membership extends CRM_Member_DAO_Membership
             }
 
             if (!$status) {
-                $fields = array( '' => array( 'title' => ts('- do not import -') ) );
+                $fields = array( '' => array( 'title' => '- ' . ts('do not import') . ' -' ) );
             } else {
-                $fields = array( '' => array( 'title' => ts('- Membership Fields -') ) );
+                $fields = array( '' => array( 'title' => '- ' . ts('Membership Fields') . ' -' ) );
             }
             
             $tmpFields     = CRM_Member_DAO_Membership::import( );
@@ -771,7 +814,7 @@ class CRM_Member_BAO_Membership extends CRM_Member_DAO_Membership
                 foreach ( $fieldsArray as $value) {
                     $tmpContactField[trim($value)] = CRM_Utils_Array::value(trim($value),$contactFields);
                     if (!$status) {
-                        $title = $tmpContactField[trim($value)]['title']." (match to contact)" ;
+                        $title = $tmpContactField[trim($value)]['title']." " . ts("(match to contact)") ;
                     } else {
                         $title = $tmpContactField[trim($value)]['title'];
                     }
@@ -779,9 +822,9 @@ class CRM_Member_BAO_Membership extends CRM_Member_DAO_Membership
                 }
             }
             $tmpContactField['external_identifier'] = $contactFields['external_identifier'];
-            $tmpContactField['external_identifier']['title'] = $contactFields['external_identifier']['title'] . " (match to contact)";
+            $tmpContactField['external_identifier']['title'] = $contactFields['external_identifier']['title'] . " " .  ts("(match to contact)");
                        
-            $tmpFields['membership_contact_id']['title'] = $tmpFields['membership_contact_id']['title'] . " (match to contact)";
+            $tmpFields['membership_contact_id']['title'] = $tmpFields['membership_contact_id']['title'] . " " .  ts("(match to contact)");;
            
             $fields = array_merge($fields, $tmpContactField);
             $fields = array_merge($fields, $tmpFields);
@@ -802,8 +845,12 @@ class CRM_Member_BAO_Membership extends CRM_Member_DAO_Membership
         $expFieldsMemType   = CRM_Member_DAO_MembershipType::export( );
         $fields = array_merge($expFieldMembership, $expFieldsMemType);
         $fields = array_merge($fields, $expFieldMembership );
+        $membershipStatus = array( 'membership_status' => array( 'title'     => 'Membership Status',
+                                                                 'name'      => 'membership_status',
+                                                                 'type'      => CRM_Utils_Type::T_STRING,
+                                                                 'where'     => 'civicrm_membership_status.name' ) );
         //CRM-6161 fix for customdata export
-        $fields = array_merge($fields, CRM_Core_BAO_CustomField::getFieldsForImport('Membership'));
+        $fields = array_merge($fields, $membershipStatus, CRM_Core_BAO_CustomField::getFieldsForImport('Membership'));
         return $fields;
     }
 
@@ -985,7 +1032,7 @@ AND civicrm_membership.is_test = %2";
             $result = null;
             if ($form->_values['is_monetary'] && !$form->_params['is_pay_later']) {
                 require_once 'CRM/Core/Payment.php';
-                $payment =& CRM_Core_Payment::singleton( $form->_mode, 'Contribute', $form->_paymentProcessor, $form );
+                $payment =& CRM_Core_Payment::singleton( $form->_mode, $form->_paymentProcessor, $form );
                 
                 if ( $form->_contributeMode == 'express' ) {
                     $result =& $payment->doExpressCheckout( $tempParams );
@@ -1085,8 +1132,8 @@ AND civicrm_membership.is_test = %2";
             if ( $form->_values['is_monetary'] && $form->_amount > 0.0 && !$form->_params['is_pay_later'] ) {
                 // this does not return
                 require_once 'CRM/Core/Payment.php';
-                $payment =& CRM_Core_Payment::singleton( $form->_mode, 'Contribute', $form->_paymentProcessor, $form );
-                $payment->doTransferCheckout( $form->_params );
+                $payment =& CRM_Core_Payment::singleton( $form->_mode, $form->_paymentProcessor, $form );
+                $payment->doTransferCheckout( $form->_params, 'contribute' );
             }
         }
 
@@ -1430,16 +1477,18 @@ SELECT c.contribution_page_id as pageID
      */
     static function deleteRelatedMemberships( $ownerMembershipId, $contactId = null ) 
     {
+        if ( !$ownerMembershipId && !$contactId ) return;  
+        
         $membership = new CRM_Member_DAO_Membership( );
         $membership->owner_membership_id = $ownerMembershipId;
-
+        
         if ( $contactId ) {
             $membership->contact_id      = $contactId;
         }
         
         $membership->find( );
         while ( $membership->fetch( ) ) {
-            // call delete function recursively since we need to delete inherited memberships of inherited memberships
+            //delete related first and then delete parent.
             self::deleteRelatedMemberships(  $membership->id );
             self::deleteMembership( $membership->id );
         }
@@ -1549,18 +1598,32 @@ WHERE  civicrm_membership.contact_id = civicrm_contact.id
                                                                         $membership->contact_id,
                                                                         CRM_Utils_Array::value( 'action', $params ) );
         }
-
+        
         // check for loops. CRM-4213
         // remove repeated related contacts, which already inherited membership.
         $relatedContactIds[$membership->contact_id] = true;
         foreach( $allRelatedContacts as $cid => $status ) {
             if ( !CRM_Utils_Array::value( $cid, $relatedContactIds ) ) {
-                $relatedContacts[$cid] =  $status;
                 $relatedContactIds[$cid] = true;
+                
+                //don't create membership again for owner contact.
+                $nestedRelationship = false;
+                if ( $membership->owner_membership_id ) {
+                    $nestedRelMembership = new CRM_Member_DAO_Membership( );
+                    $nestedRelMembership->id = $membership->owner_membership_id;
+                    $nestedRelMembership->contact_id = $cid;
+                    $nestedRelationship = $nestedRelMembership->find( true );
+                    $nestedRelMembership->free( );
+                }
+                if ( !$nestedRelationship ) $relatedContacts[$cid] =  $status;
             }
         }
         
-        if ( ! empty($relatedContacts) ) {
+        //lets cleanup related membership if any.
+        if ( empty( $relatedContacts ) ) {
+            require_once 'CRM/Member/BAO/Membership.php';
+            CRM_Member_BAO_Membership::deleteRelatedMemberships( $membership->id );
+        } else {
             // Edit the params array
             unset( $params['id'] );
             // Reminder should be sent only to the direct membership
@@ -1614,11 +1677,8 @@ WHERE  civicrm_membership.contact_id = civicrm_contact.id
                     }
                 }
                 
-                //avoid extra processing.
-                if ( $relMembership->id && 
-                     $params['status_id'] == $relMembership->status_id ) {
-                    continue;
-                }
+                //don't calculate status again in create( );
+                $params['skipStatusCal'] = true;
                 
                 //do create activity if we changed status. 
                 if ( $params['status_id'] != $relMembership->status_id ) {
@@ -1711,4 +1771,5 @@ FROM   civicrm_membership_type
         $query = "SELECT count(*) FROM civicrm_membership WHERE civicrm_membership.contact_id = {$contactID} AND civicrm_membership.is_test = 0 ";
         return CRM_Core_DAO::singleValueQuery( $query );
     }
+   
 }
